% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weight.R
\name{calculate_weight}
\alias{calculate_weight}
\title{Calculated empirical density and weight based on variable split.}
\usage{
calculate_weight(profiles, data, variable_split)
}
\arguments{
\item{profiles}{data.frame generated by ingredients::ceteris_paribus()}

\item{data}{data.frame with raw data to model}

\item{variable_split}{list generated by ingredients::calculate_variable_split()}
}
\value{
Return an weight based on empirical density.
}
\description{
This function calculate an empirical density of raw data based on variable split from Ceteris Paribus profiles. Then calculated weight for values generated by ingredients::ceteris_paribus().
}
\examples{
\dontrun{

split <- ingredients::calculate_variable_split(data, variables = colnames(data))
calculate_weight(profiles, data, variable_split = split)
}

}
