% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEmpiricalDensity.R
\name{CalculateEmpiricalDensity}
\alias{CalculateEmpiricalDensity}
\title{Calculated empirical density.}
\usage{
CalculateEmpiricalDensity(cp, df, variable_split)
}
\arguments{
\item{cp}{data.frame generated by ingredients::ceteris_paribus()}

\item{df}{data.frame with raw data to model}

\item{variable_split}{list generated by ingredients::calculate_variable_split()}
}
\value{
Return an empirical density based on ingredients::calculate_variable_split().
}
\description{
This function calculate an empirical density based on variable split from Ceteris Paribus profiles.
}
\examples{
\dontrun{

split <- ingredients::calculate_variable_split(df, variables = colnames(df))
CalculateEmpiricalDensity(df, variable_split = split)
}

}
