% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-stand-alone.R
\name{launch_gizmo_local}
\alias{launch_gizmo_local}
\title{launch_gizmo_local}
\usage{
launch_gizmo_local(gizmo, gizmo_name, state = NULL,
  web_browser = paneViewer())
}
\description{
launch_gizmo_local
}
