% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_vivid.R
\name{launch_vivid}
\alias{launch_vivid}
\title{launch_vivid}
\usage{
launch_vivid(launch.browser, ...)
}
\arguments{
\item{launch.browser}{passed to \code{shiny::runApp}}

\item{...}{passed to \code{shiny::runApp}}
}
\description{
Second entry point for vivid package
}
