% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalVariableImportanceViaOscillations.R
\name{LocalVariableImportanceViaOscillations}
\alias{LocalVariableImportanceViaOscillations}
\title{Local Variable Importance measure based on Ceteris Paribus profiles.}
\usage{
LocalVariableImportanceViaOscillations(cp, absolute_deviation = TRUE,
  point = TRUE, density = FALSE)
}
\arguments{
\item{cp}{data.frame generate by ceterisParibus::ceteris_paribus()}

\item{absolute_deviation}{logical parameter, if `absolute_deviation = TRUE` then measue is calculated as absolute deviation, else is calculated as a root from average squares}

\item{point}{logical parameter, if `point = TRUE` then measure is calculated as a distance from f(x), else measure is calculated as a distance from average CP}

\item{density}{logical parameter, if `density = TRUE` then measure is weighted based on the density of variable, else is not weighted}
}
\description{
This function calculate local importance measure in eight variants.
}
\examples{
\dontrun{

LocalVariableImportanceViaOscillations(cp, absolute_deviation = TRUE, point = TRUE, density = FALSE)

LocalVariableImportanceViaOscillations(cp, absolute_deviation = TRUE, point = FALSE, density = FALSE)
}

}
