% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.local_importance}
\alias{plot.local_importance}
\title{Plots Local Variable Importance measure}
\usage{
\method{plot}{local_importance}(measure)
}
\arguments{
\item{measure}{object returned from LocalVariableImportanceViaOscillations}
}
\description{
Function plot.local_importance plots local importance measure based on Ceteris Paribus profiles.
}
\examples{
\dontrun{
measure <- LocalVariableImportanceViaOscillations(cp, df, absolute_deviation = TRUE, point = TRUE, density = FALSE, kernel_density = "gaussian", bw_density = "nrd0")
plot(measure)
}

}
