% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Local Variable Importance measure based on Ceteris Paribus profiles.}
\usage{
plot(cp, df)
}
\arguments{
\item{cp}{data.frame generate by ceterisParibus::ceteris_paribus()}

\item{df}{data.frame with raw data to model}
}
\description{
This function plot local importance measure in eight variants.
}
\examples{
\dontrun{
plot(cp, df)
}

}
