% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-exec.R
\docType{data}
\name{QueueLinkedR}
\alias{QueueLinkedR}
\title{QueueLinkedR}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueueLinkedR
}
\description{
QueueLinkedR
}
\keyword{datasets}
