% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-input-number.R
\name{run_gizmo_input_number}
\alias{run_gizmo_input_number}
\title{run_gizmo_input_number with run_standalone}
\usage{
run_gizmo_input_number()
}
\description{
run_gizmo_input_number with run_standalone
}
