% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-input-vector.r
\name{run_gizmo_input_vector}
\alias{run_gizmo_input_vector}
\title{run_gizmo_input_vector with run_standalone}
\usage{
run_gizmo_input_vector()
}
\description{
run_gizmo_input_vector with run_standalone
}
