% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-load-package-data.R
\name{run_gizmo_load_package_data}
\alias{run_gizmo_load_package_data}
\title{run_gizmo_load_package_data with run_standalone}
\usage{
run_gizmo_load_package_data()
}
\description{
run_gizmo_load_package_data with run_standalone
}
