% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-plot-barplot.R
\name{run_gizmo_plot_barplot}
\alias{run_gizmo_plot_barplot}
\title{run_gizmo_plot_barplot with run_standalone}
\usage{
run_gizmo_plot_barplot()
}
\description{
run_gizmo_plot_barplot with run_standalone
}
