% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-plot-boxplot.R
\name{run_gizmo_plot_boxplot}
\alias{run_gizmo_plot_boxplot}
\title{run_gizmo_plot_boxplot with run_standalone}
\usage{
run_gizmo_plot_boxplot()
}
\description{
run_gizmo_plot_boxplot with run_standalone
}
