% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-plot-dotplot.R
\name{run_gizmo_plot_dotchart}
\alias{run_gizmo_plot_dotchart}
\title{run_gizmo_plot_dotchart with run_standalone}
\usage{
run_gizmo_plot_dotchart()
}
\description{
run_gizmo_plot_dotchart with run_standalone
}
