% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-plot-pointslines.R
\name{run_gizmo_plot_pointslines}
\alias{run_gizmo_plot_pointslines}
\title{run_gizmo_plot_pointslines with run_standalone}
\usage{
run_gizmo_plot_pointslines()
}
\description{
run_gizmo_plot_pointslines with run_standalone
}
