% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo-plot-scatter.R
\name{run_gizmo_plot_scatter}
\alias{run_gizmo_plot_scatter}
\title{run_gizmo_plot_scatter with run_standalone}
\usage{
run_gizmo_plot_scatter()
}
\description{
run_gizmo_plot_scatter with run_standalone
}
