% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid.R
\name{vivid}
\alias{vivid}
\title{vivid}
\usage{
vivid(child_process = FALSE, launch.browser = TRUE, ...)
}
\arguments{
\item{child_process}{\code{TRUE} for production mode; \code{FALSE} for debug mode}

\item{launch.browser}{passed to \code{launch_vivid} to \code{shiny::runApp}}
}
\description{
Entry point for vivid package
}
