% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalVariableImportanceViaOscillations.R
\name{LocalVariableImportanceViaOscillations}
\alias{LocalVariableImportanceViaOscillations}
\title{Local Variable Importance measure based on Ceteris Paribus profiles.}
\usage{
LocalVariableImportanceViaOscillations(cp, df, absolute_deviation = TRUE,
  point = TRUE, density = TRUE)
}
\arguments{
\item{cp}{data.frame generated by ingredients::ceteris_paribus()}

\item{df}{data.frame with raw data to model}

\item{absolute_deviation}{logical parameter, if `absolute_deviation = TRUE` then measue is calculated as absolute deviation, else is calculated as a root from average squares}

\item{point}{logical parameter, if `point = TRUE` then measure is calculated as a distance from f(x), else measure is calculated as a distance from average CP}

\item{density}{logical parameter, if `density = TRUE` then measure is weighted based on the density of variable, else is not weighted}
}
\value{
A list of the class 'local_variable_importance'.
It's a list with calculated local variable importance measure.
}
\description{
This function calculate local importance measure in eight variants.
}
\examples{
\dontrun{

LocalVariableImportanceViaOscillations(cp, df, absolute_deviation = TRUE, point = TRUE, density = FALSE, kernel_density = "gaussian", bw_density = "nrd0")

LocalVariableImportanceViaOscillations(cp, df, absolute_deviation = TRUE, point = TRUE, density = FALSE, kernel_density = "gaussian", bw_density = "SJ")
}

}
