% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_local_importance.R
\name{plot.local_importance}
\alias{plot.local_importance}
\title{Plot Local Variable Importance measure}
\usage{
\method{plot}{local_importance}(x, ...)
}
\arguments{
\item{x}{object returned from `local_variable_importance()` function}

\item{...}{other parameters}
}
\value{
a ggplot2 object
}
\description{
Function plot.local_importance plots local importance measure based on Ceteris Paribus profiles.
}
\examples{
\dontrun{
library("vivo", warn.conflicts = FALSE, quietly = TRUE)
measure <- local_variable_importance(profiles, data,
                                     absolute_deviation = TRUE, point = TRUE, density = FALSE)
plot(measure)
}

}
